requirements.txt
annotated-types==0.7.0
anyio==4.4.0
certifi==2024.8.30
charset-normalizer==3.3.2
colorama==0.4.6
distro==1.9.0
filelock==3.16.0
fsspec==2024.9.0
gitdb==4.0.11
GitPython==3.1.41
h11==0.14.0
httpcore==1.0.5
httpx==0.27.2
huggingface-hub==0.24.6
idna==3.8
inflect==7.2.1
jiter==0.5.0
more-itertools==10.4.0
neo4j==5.17.0
packaging==24.1
psycopg2-binary==2.9.9
pydantic==2.9.1
pydantic_core==2.23.3
python-dotenv==1.0.1
pytz==2024.1
PyYAML==6.0.2
requests==2.32.3
smmap==5.0.1
sniffio==1.3.1
tqdm==4.66.5
typeguard==4.3.0
typing_extensions==4.12.2
urllib3==2.2.2
